Convert torsion drive job and result fetching to track ID
The torsion drive job and results tracking now searches and tracks
by ID instead of Hash. Partial progress to #60.

Outstanding questions: In the `submit_optimization_tasks` call, the
check for `completed` jobs will return the ID of the *procedure_id*, but
the rest of the jobs are based on the *task_id*. This is a bug, however,
none of the tests reach this code and none of our own work has tripped
it since #57 introduced a syntax bug which would have been raised
if it ever did. We should try to fix this, but I don't know of a good
test to reach this code outside of production.