Ensure torsion drive completed tasks pull from their queue id
This makes sure the `job_map` for new Torsion Drive service tasks
reference the same table in the storage layer rather than different
layers.

A new bug was identified for a rare corner case where a secondary
service submits jobs after the initial query from a primary service, but
before the tasks from the primary are submitted to the queue. We have
elected not to fix that in this change, but will raise a helpful error
if someone ever actually gets it.

Fixes #62