Work around missing SGECluster job_queue
This PR adds a fix to the SGECluster which was missing a job_queue
kwarg. A formal fix is proposed in dask/dask-jobqueue\#256.

For now, there is a check that should `job_queue` ever get patched in,
we will fall back to the provided one so updating dask-jobqueue does not
require updating Fractal as well