Stale Job Tracking
This PR introduces better handling and logging for failed job
`post` events. Failed `update` actions are queued and retried to a
set number of times before the manager drops the jobs and considers
them "stale." A maximum number of stale updates are allowed to be
tracked before the server just shuts itself down.

I'm working on a test for this block of code, but its ready otherwise

This also fixes a bug in the `dask` queue_manager CLI script where
`env_extra` was misspelled and thus ignored by `dask`. (it was
`env_exta`).