Manager settings QOL Update: Names
Introduces some name changes to some of the Manager options to be
more exact on what they do. The following changes are proposed:

* `ntasks` -> `tasks_per_job`
* `ncores` -> `cores_per_job`
* `memory` -> memory_per_job`
* `max_nodes` -> max_cluster_jobs`
    * This one is already in but open for discussion since no release tied to it yet

This also changes `max_tasks` to be functionally set by default as:

`ceil(cores_per_job * max_cluster_jobs * 1.2) + 1`

Users can still set this, but are discouraged from doing so.

Opinions on this (and any setting I didn't touch too!)