Make a Doc continuity pass for 0.7
I've gone though and touched many of the pages to keep the docs in line
with one another as far as features, content, and references. I'm sure
there are plenty of missed items, but I think this will be a good step.

I also updated the `devtools` README.md file to be accurate to the
folder contents and how we go about things.