Remove dask-jobqueue workarounds
This also now adds an implicit minimum version pin to Dask Jobqueue
0.5.0 and Parsl 0.8.0 should users call those.

I also went back and fixed the dask node exclusivity setting for the
other clusters to the best of my ability and the docs I could find.
I improved the docs for that setting to help users should it throw an
error.