Add Server CLI Documentation
This adds the `qcfractal-server` CLU documentation for all its
sub-options: init, start, config, and upgrade. This included using
the auto-doc wrapper for Pydantic `Schema`-rich classes, which had to
be moved higher up in the directory structure for easy imports. If need
be, I can move it back down into the CLI Utils instead of top level
Utils and import from there.

There are also a number of other fixes found in writing these listed
below:

* Fix Sphinx errors about missing `_static` directory
* Removed doc reference to a `database_design.jpg` file which was removed a while ago
* Renamed arg `--disable_ssl` to `--disable-ssl` to match other args (CLI hyphens vs underscore)
* Fixed bug in Parsl Adapter worker efficiency raising when it should not have
* Small typos in docstring fixes
* Fixed logger bug in PostgresHarness