[WIP] Add Schema descriptions to all Pydantic models in Fractal
As it says in the title. Will be helpful for auto-doc generation
and future developers.

Still a work in progress. There is alot here though so I wanted to post
what I have as I keep editing. There are a few TODOs I have added
that I think need extra attention or that I have questions on.

Due to the large changes and the need for accuracy, I could use
some eyes on each of these to make sure I have described
each option correctly.