[json-591] Conceptual split of *_json into *_json, *_dict methods on Collections
Addresses #591

Pydantic models support direct JSON serialization; we take advantage of
this here to support use of e.g. sets in a model.

The `to_json` and `from_json` methods also technically dealt with dicts,
not JSON strings; this is a distinction that matters since not all dicts
are valid JSON constructs. We resolve that distinction here.

Need explicit roundtripping tests for `*_json` and `*_dict` methods yet.