Handle orphan kv_store better
For typical procedures, we add to kv_store, then delete
in order to preserve atomicity.

For services, we can upsert to kv_store. This is because the output/error
is still stored in the service object, and would be re-commited to the database
at a later time if there is an error.