`BaseProcedureDataset` now has refactored `status` method
In our effort to make the `OptimizationDataset` object more stateless in
its delivery of calculation results, which can change as compute occurs,
`status` now calls the cache and always grabs what it doesn't have from
the server; it no longer holds onto ERRORed and INCOMPLETE results in a
`pandas.DataFrame`.

This `status` method now has the sole responsibility of reporting *status*.
Its display/returns can be also altered with a variety of keyward args.
This allows it to work by default as a human-in-the-loop informational
method, but also as a programmatic method with parsable outputs.

Because `qcfractal` requires `pandas` anyway, we also include an `asdf`
keyword that reproduces the existing behavior of returning a DataFrame.