Small shim in `FractalClient.query_results`
There are some datasets in the database with old QCElemental result models that
fail validation when queries. These are retruned by
`FractalClient._automodel_request` as dictionaries, even without
`include` specified. This shim handles these cases as gracefully as we
can at this time until the old results are migrated to the new result
model.